pyproject.toml
[project]
name = "genealogytree"
version = "0.1.0"
description = ""
authors = [
    {name = "arturgorski14",email = "arturgorski14@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "neo4j (>=5.28.1,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"
pre-commit = "^4.2.0"
pytest-depends = "^1.0.1"
httpx = "^0.28.1"
ruff = "^0.12.10"
pytest-cov = "^6.2.1"

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88